Damage: Both the human and monsters so "random" damage based on their strength. I personally don't like things being so predictable like this enemy need three hits to kill, etc. so this is good but if the group doesn't like it we can get rid of it. Also when defending we don't want to block ALL damage. So now it cuts the random damage value in half and if it's lower than 5 it does nothing. Otherwise it does the damage after it was cut in half.
Items: I tried to make an item class or modify a monster to be an item but I have no idea how to make images/animations. So now our only item is the potion which looks like an attacking plus sign. It heals 50 hp and 50 mana. We need to decide how much is the max hp/mana and if this is variable in any way.

Sounds: They(it) work(s) :D

Bugs I found: after an enemy attacks and you move away, the enemy stays in it's neutral position even when it is walking towards you again.