Added a boss class that shouldn't move because I'm assuming these will be massive monsters that take up most of the screen. We'll see when the images are up. There's a problem I can't fix though. I dunno why the boss doesn't want to move directly to the edge of the Hero's walking area. See line 207 in level1.py
I want to add a damage or modify the attack function in the player class because it is already copy-pasted three times in the level1.py. I'll do it later tonight.

Also I know why the jumping doesn't work but I can't think of a solution that will fix it. The problem is that the jump animation is twice the height of the other animations so everything gets pushed down. I also don't know how we will be able to have multiple actions in the jump or if it will have any use at all since you will still get hit from the ground.