YAY FOR MULTIPLE ENEMIES.
Stuff I changed: I had to lower the chance of an enemy spawning when you move right because there were times when I had two or more pop up at once. Also the P key spawns the weakest enemy in the lower-left of the screen. For testing, etc. Careful: in large groups they do massive damage.

Had to get rid of the stats in the upper right corner :( I don't think players will like not being able to see how much more health enemies have but I dunno any good way to fix this right now....

Stuff to do: Set a limit for the number of enemies. Unless you like lag this is pretty important. Also we don't want players to be at 9 kills then to kill two enemies meaning they skip the boss. We also have to figure out when we want to really spawn the boss. And we have to make the boss.