Now able to spawn enemies using "p" in level3
Found some bugs related to the boss. See line 80 in level.py

Displays stats of the boss when and only when it is there. Will show stats of the latest enemy to spawn if you hit p to spawn that guy in the corner.

Also no more than 9 enemies on screen at once.